+++ .eslintignore
@@ -1,0 +1,2 @@
+models/index.js
+node_modules
