+++ .eslintrc.json
@@ -1,0 +1,40 @@
+{
+  "parserOptions": {
+    "ecmaVersion": 2017
+  },
+  "env": {
+    "es6": true,
+    "browser": true,
+    "commonjs": true,
+    "node": true
+  },
+  "extends": "prettier",
+  "plugins": ["prettier"],
+  "rules": {
+    "array-callback-return": "error",
+    "camelcase": "error",
+    "curly": "error",
+    "dot-notation": "error",
+    "eqeqeq": "error",
+    "implicit-arrow-linebreak": ["error", "beside"],
+    "indent": ["error", 2],
+    "no-duplicate-case": "error",
+    "no-else-return": "error",
+    "no-empty": "error",
+    "no-empty-function": "error",
+    "no-extra-semi": "error",
+    "no-func-assign": "error",
+    "no-irregular-whitespace": "error",
+    "no-multi-spaces": "error",
+    "no-unreachable": "error",
+    "no-unused-vars": "error",
+    "no-var": "error",
+    "prefer-arrow-callback": "error",
+    "prefer-const": "error",
+    "prettier/prettier": "error",
+    "quotes": ["error", "double"],
+    "semi": ["error", "always"],
+    "vars-on-top": "error",
+    "yoda": "error"
+  }
+}
