+++ .travis.yml
@@ -1,0 +1,12 @@
+language: node_js
+node_js: "12"
+branches:
+  only:
+    - master
+cache:
+  directories:
+    - node_modules
+services:
+  - mysql
+before_install:
+  - mysql -e 'CREATE DATABASE IF NOT EXISTS testdb;'
