--- Sequelize-Passport-Example/.gitignore
@@ -1,3 +1,0 @@
-node_modules/
-.env
-.DS_Store
