--- Sequelize-Passport-Example/.travis.yml
@@ -1,12 +1,0 @@
-language: node_js
-node_js: "12"
-branches:
-  only:
-    - master
-cache:
-  directories:
-    - node_modules
-services:
-  - mysql
-before_install:
-  - mysql -e 'CREATE DATABASE IF NOT EXISTS testdb;'
