--- Sequelize-Passport-Example/db/rating.sql
@@ -1,8 +1,0 @@
-USE_db;
-
-INSERT INTO Rating (Score,ratee_id,rater_id)
-    VALUES (1,'hema2','hema123'),
-           (1,'hema2','hema123'),
-           (1,'hema2','hema123')
-           (1,'hema2','hema123')
-           (1,'hema2','hema123')
