--- Sequelize-Passport-Example/db/wishlist.sql
@@ -1,9 +1,0 @@
-USE books_db;
-
--- wishlist--
-INSERT INTO  wishlist(id,user_id,wishlist_id,name)
-Values(1,'hema21',abcdef,hemawishlist),
-      (2,'hema21',abcdef,hemawishlist),
-      (3,'hema21',abcdef,hemawishlist),
-      (4'hema21',abcdef,hemawishlist),
-      (5'hema21',abcdef,hemawishlist)
