+++ db/customers.sql
@@ -1,0 +1,12 @@
+SELECT * FROM books_db.customer;
+
+USE books_db;
+
+INSERT INTO Customer (login_id,name,password,phone_num)
+            VALUES ('1','George','789','12345');
+INSERT INTO Customer (login_id,name,password,phone_num)
+            VALUES ('2','oscar','456','12345');
+INSERT INTO Customer (login_id,name,password,phone_num)
+            VALUES ('3','hema','123','12345');
+INSERT INTO Customer (login_id,name,password,phone_num)
+            VALUES ('4','Joe','678','12345')
