+++ db/wishlist.sql
@@ -1,0 +1,12 @@
+USE books_db;
+
+-- wishlist--
+INSERT INTO wishlist (user_id,wishlist_id)
+ Values('1', '345');
+INSERT INTO  wishlist(user_id,wishlist_id)
+    values  ('2','678');
+INSERT INTO  wishlist(user_id,wishlist_id)
+     values ('3','123');
+INSERT INTO  wishlist(user_id,wishlist_id)
+     values ('4','678')
+
