+++ models/userhistory.js
@@ -1,0 +1,9 @@
+$(document).on('click', ".transact", function(){
+	var orderNumber = $(this).val();
+	if($('#' + orderNumber).is(':visible')){
+		$('#' + orderNumber).hide(1000);
+	}
+	else{
+		$('#' + orderNumber).show(1000);
+	}
+});
