+++ public/stylesheets/style.css
@@ -1,0 +1,4 @@
+form.signup,
+form.login {
+  margin-top: 50px;
+}
